new script commands and misc

+++ Utilities/SEMUtilities.cpp
@@ -531,7 +531,7 @@ char UtilCheckIllegalChars(CString &filename, int slashOrDriveOK, CString descri
 void UtilSplitString(CString fullStr, const char *delim, 
   std::vector<std::string> &strList)
 {
-  int ind, lenDelim = strlen(delim);
+  int ind, lenDelim = (int)strlen(delim);
   strList.clear();
   while ((ind = fullStr.Find(delim)) >= 0) {
     strList.push_back((LPCTSTR)fullStr.Left(ind));
